saving country details in db

+++ src/main/java/com/laan/sportsda/client/response/CountryClientResponse.java
@@ -14,7 +14,7 @@ public class CountryClientResponse {
     private String cca3;
     private String ccn3;
     private List<String> capital;
-    private Long area;
+    private Double area;
     private Long population;
     private String region;
     private String subregion;
